requirements/base.txt
pytz==2021.1  # https://github.com/stub42/pytz
python-slugify==5.0.2  # https://github.com/un33k/python-slugify
Pillow==8.2.0  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==1.1.0  # https://github.com/redis/hiredis-py
celery==4.4.6  # pyup: < 5.0,!=4.4.7  # https://github.com/celery/celery
django-celery-beat==2.2.0  # https://github.com/celery/django-celery-beat
flower==0.9.7  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==3.1.12  # pyup: < 3.2  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.44.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.12.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-compressor==2.4.1  # https://github.com/django-compressor/django-compressor
django-redis==5.0.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.12.4  # https://github.com/encode/django-rest-framework
django-cors-headers==3.7.0 # https://github.com/adamchainz/django-cors-headers



# Admin
# ------------------------------------------------------------------------------
django-jet2==1.0.12
FeedParser==5.2.1
django-ipware==3.0.1
django-autocomplete-light==3.8.2
# django-flatpages-tinymce==0.1.1
django-tinymce==3.3.0
django-filebrowser-no-grappelli==3.8.0
google-api-python-client==1.4.1
django-admin-honeypot==1.1.0
django-cookie-law==2.0.5  # https://pypi.org/project/django-cookie-law/
django-cookie-consent==0.2.6 # https://pypi.org/project/django-cookie-consent/

# Fileifeld upload validator
python-magic==0.4.24


# Minify HTML in django
# ------------------------------------------------------------------------------
django-shell-plus==1.1.7

# Minify HTML in django
# ------------------------------------------------------------------------------
# Todo: find the uses of these to minify wesbites and others to improve security
django-htmlmin==0.11.0
python-memcached==1.59


# API NECESSARY APPS
# ------------------------------------------------------------------------------
djangorestframework-simplejwt==4.6.0 # Add proper token creation url links for app
drf-yasg==1.20.0 # Add documentation for api

# Add swagger for friendlier api view
# ------------------------------------------------------------------------------
django-rest-swagger==2.2.0

# Documentation API
# ------------------------------------------------------------------------------
coreapi==2.3.3 # https://github.com/core-api/python-client


# PHONE NUMBER VERIFICATION
# ------------------------------------------------------------------------------
django-phonenumber-field==5.0.0
phonenumbers==8.11.1


# List of countries
# ------------------------------------------------------------------------------
django-countries-plus==1.3.1


# Add image resize for models
# ------------------------------------------------------------------------------
django-resized==0.3.11


# Add geolocation to django app
# ------------------------------------------------------------------------------
geoip2==4.1.0
maxminddb==2.0.3


# Add medium type of read time to a post
# ------------------------------------------------------------------------------
readtime==1.1.1

# Dynamic Scraper for django
# ------------------------------------------------------------------------------
# Scrapy is a fast high-level web crawling and web scraping framework, 
# used to crawl websites and extract structured data from their pages. 
# It can be used for a wide range of purposes, from data mining to 
# monitoring and automated testing.
# ------------------------------------------------------------------------------
webdriver-manager==3.4.2
lxml==4.6.3 
scrapy-djangoitem==1.1.1
word2word==1.0.0 
requests-html==0.10.0
django-filter==2.4.0 

#Todo: Django Filter Search, Ajax Commenting, Language Integration, Cookie Consent
